refactor(repository): reorganiza repositórios por domínio
