+++ src/main/java/com/project/miinhareceita/services/AuthService.java
@@ -1,7 +1,7 @@
 package com.project.miinhareceita.services;
 
-import com.project.miinhareceita.repositories.UserRepository;
 import com.project.miinhareceita.user.domain.User;
+import com.project.miinhareceita.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
