+++ src/main/java/com/project/miinhareceita/services/FavoriteService.java
@@ -3,9 +3,9 @@ package com.project.miinhareceita.services;
 import com.project.miinhareceita.dtos.FavoriteDTO;
 import com.project.miinhareceita.dtos.FavoriteInsertDTO;
 import com.project.miinhareceita.favorite.domain.Favorite;
+import com.project.miinhareceita.favorite.repository.FavoriteRepository;
 import com.project.miinhareceita.recipe.domain.Recipe;
-import com.project.miinhareceita.repositories.FavoriteRepository;
-import com.project.miinhareceita.repositories.RecipeRepository;
+import com.project.miinhareceita.recipe.repository.RecipeRepository;
 import com.project.miinhareceita.services.exceptions.DatabaseException;
 import com.project.miinhareceita.services.exceptions.ForbiddenException;
 import com.project.miinhareceita.services.exceptions.ResourceNotFoundException;
