+++ src/main/java/com/project/miinhareceita/services/RecipeService.java
@@ -4,12 +4,12 @@ import com.project.miinhareceita.dtos.RecipeDTO;
 import com.project.miinhareceita.dtos.RecipeIngredientsDTO;
 import com.project.miinhareceita.dtos.RecipeMinDTO;
 import com.project.miinhareceita.ingredient.domain.Ingredients;
+import com.project.miinhareceita.ingredient.repository.IngredientsRepository;
 import com.project.miinhareceita.projections.RecipeProjections;
 import com.project.miinhareceita.recipe.domain.Recipe;
 import com.project.miinhareceita.recipe.domain.RecipeIngredients;
-import com.project.miinhareceita.repositories.IngredientsRepository;
-import com.project.miinhareceita.repositories.RecipeIngredientsRepository;
-import com.project.miinhareceita.repositories.RecipeRepository;
+import com.project.miinhareceita.recipe.repository.RecipeIngredientsRepository;
+import com.project.miinhareceita.recipe.repository.RecipeRepository;
 import com.project.miinhareceita.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
