+++ src/main/java/com/project/miinhareceita/services/ReviewService.java
@@ -3,14 +3,14 @@ package com.project.miinhareceita.services;
 import com.project.miinhareceita.dtos.ReviewDTO;
 import com.project.miinhareceita.projections.ReviewProjections;
 import com.project.miinhareceita.recipe.domain.Recipe;
-import com.project.miinhareceita.repositories.RecipeRepository;
-import com.project.miinhareceita.repositories.ReviewRepository;
-import com.project.miinhareceita.repositories.UserRepository;
+import com.project.miinhareceita.recipe.repository.RecipeRepository;
 import com.project.miinhareceita.review.domain.Review;
+import com.project.miinhareceita.review.repository.ReviewRepository;
 import com.project.miinhareceita.services.exceptions.DatabaseException;
 import com.project.miinhareceita.services.exceptions.ForbiddenException;
 import com.project.miinhareceita.services.exceptions.ResourceNotFoundException;
 import com.project.miinhareceita.user.domain.User;
+import com.project.miinhareceita.user.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
