+++ src/main/java/com/project/miinhareceita/services/validation/UserInsertValidator.java
@@ -2,8 +2,8 @@ package com.project.miinhareceita.services.validation;
 
 import com.project.miinhareceita.controllers.exceptions.FieldMessage;
 import com.project.miinhareceita.dtos.UserInsertDTO;
-import com.project.miinhareceita.repositories.UserRepository;
 import com.project.miinhareceita.user.domain.User;
+import com.project.miinhareceita.user.repository.UserRepository;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import org.springframework.beans.factory.annotation.Autowired;
