+++ src/main/java/com/project/miinhareceita/user/repository/RoleRepository.java
@@ -1,4 +1,4 @@
-package com.project.miinhareceita.repositories;
+package com.project.miinhareceita.user.repository;
 
 import com.project.miinhareceita.user.domain.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
